Refactor Store
